requirements.txt
absl-py==0.1.10
aniso8601==2.0.0
bleach==1.5.0
click==6.7
cycler==0.10.0
enum34==1.1.6
Flask==0.12.2
Flask-RESTful==0.3.6
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
joblib==0.11
Keras==2.1.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==5.0.0
protobuf==3.5.1
pudb==2017.1.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==1.5.1
urwid==2.0.1
Werkzeug==0.14.1
